Refactor: fencer: pass result object when sending an async reply
... via stonith_send_async_reply(), instead of sending the mapped legacy code
and action stdout separately. Also, drop the "stonith_" prefix since the
function is static.

This moves the mapping from the stonith_send_async_reply() callers to the
function itself, so we use the result object and standard codes as long as
possible, and map to a legacy code only where needed.