Fix: libcrmservice: properly detect systemd unit existence
Previously, we treated a successful LoadUnit method call as indicating that a
systemd unit exists. However, it does not mean the unit was successfully
loaded, only that the LoadUnit got a reply of the expected type. We must also
check the LoadState property to know whether the load itself was successful.