Test: cts-exec: exclude nagios from tests if present
Apparently, no one has previously tried to run the executor regression tests on
a node with nagios resource meta-data installed, which speaks to the feature's
lack of use.

Ideally, we'd create a dummy nagios check and add some tests for it, but given
the lack of interest in the feature, it's not worth the effort at this time.