Low: libcrmservice: improve exit codes for child execution failures
After a child process for a file-based action has been forked but before it has
executed the file, a few things can go wrong. Previously, the process exited
PCMK_OCF_NOT_CONFIGURED for all these.

Now, functionize getting a standard-appropriate exit code for "not configured",
and use that when unable to get CIB secrets. Use a standard-appropriate
"unauthorized" error if unable to set group or user.