Fix: daemons: Re-register UID/GID with corosync on restart.
If the corosync daemon crashes and then is restarted, we need to
re-register the UID/GID with it so the subdaemons are allowed to connect
to corosync.  Otherwise, they can fail with an EACCES.

Additionally, change the return codes to G_SOURCE_{REMOVE,CONTINUE}.
Returning _REMOVE on success is new to this function, but appears to fix
a segfault from glib when corosync restarts.  It appears something about
the restart process causes the timer callback struct to be invalid but
not NULL, which results in a segfault.  Once we have successfully
reconnected, we don't need this callback around.