Refactor: daemons: Add a function to see if a cluster is available.
The generic pcmkd_cluster_connected function will call a
cluster-specific function or simply return true if no cluster is
supported.  Right now, we only support corosync so there's only
pcmkd_corosync_connected.

This function attempts to initialize and then shut down a connection to
the corosync CPG API.  If that is successful, we know that corosync is
available to use.