Fix: fencer: handle dynamic target query failures better
Previously, the callbacks for list and status queries checked only the result's
exit status. However, the services library will use PCMK_OCF_UNKNOWN_ERROR (1)
as the exit status for internal failures, and that value signifies a recognized
node (not an error) for fence list actions.

Now, the callbacks check the execution status as well.