Feature: daemons: Use formatted output for scheduler help, errors, etc.
This daemon is already using formatted output - it uses a logging
formatter to output everything it's doing to the log file.  This commit
adds a second formatter that is used only for help and version, errors,
and other interactive usage.  The metadata output has not yet been
worked into using this formatter.