Build: rpm: allow specifying an arbitrary directory for RPM artifacts
The rpm targets already support RPMDEST of "subtree" or "toplevel" to determine
where to put RPM artifacts such as spec files and source RPMs.

Now, RPMDEST can also be set to any arbitrary path, and all RPM artifacts will
be put in that location. This can be useful when running mock manually instead
of using the mock-related targets, for example (without running automake):

	mkdir -p $DISTRO
	make -C rpm -f Makefile.am RPMDEST=$PWD/$DISTRO/rpm srpm
	mock --no-clean --root $DISTRO --resultdir=$PWD/$DISTRO \
		$DISTRO/rpm/SRPMS/pacemaker*.src.rpm