Log: libcrmservice: improve messages for pre-execution failures in child
After the services library forks a child process to execute a file, only exit
status is available to communicate errors, not execution status. However, if
it's for an OCF agent, we can output an exit reason for internal failures as if
the agent did, so the parent can parse and record that for better status
displays.