Low: libcrmservice: distinguish local and global configuration errors
services__configuration_error() now takes an argument to distinguish a
configuration error that only affects the local node from one that affects the
whole cluster (corresponding to the difference between PCMK_OCF_INVALID_PARAM
and PCMK_OCF_NOT_CONFIGURED).

The lone caller, for when CIB secrets could not be loaded, now uses the
local-only error. Previously, that situation would cause the resource to be
unrunnable across the entire cluster, but it's possible only the local node is
missing the secrets due to a synchronization issue or such.