Low: Use user_include_exclude_cb in crm_mon more.
For options that the user provides on the command line that affect the
sections (things like -D), use the user_ version of the callback.  This
version is meant to override anything we may have done internally.  Note
that the display of fencing options is complicated and is basically the
reason multiple versions of this callback exist, so they are not touched
by this patch.

See: clbz#5485