Feature: libpacemaker: improve result for high-level fencing API
Previously, pcmk__fencing_action()'s helpers for asynchronous fencing actions
initialized the result to a generic error, and then overrode that only on
success.

Now, set a detailed result for early failures, and use the full result when
available from the fencing API.

A standard return code is still returned to callers at this point.