Low: fencer,scheduler: compare fence action names case-sensitively
Use the new convenience function pcmk__is_fencing_action() to check whether
an action name is a fencing action ("off", "reboot", or "poweroff"). This
changes the behavior from case-insensitive to case-sensitive, which is more
appropriate (the case-insensitivity was inherited from lazy use of the old
safe_str_eq() function which was always case-insensitive).