Low: fencer: improve broadcasting replies for fenced originators
If the target of a fencing action was also the originator, the executioner
broadcasts the result on their behalf.

Previously, it would check if the action was not in a list of actions that are
never broadcasted. However we really only want to broadcast off/reboot results
so just check for that instead.

This also rearranges reply creation slightly so we don't trace-log the reply
until it is fully created.