Refactor: fencer: add convenience function for setting protocol error result
The fencer will soon track and return the full result (rather than just a
legacy return code) for fencing actions, for callbacks and notifications.
To simplify that process as well as move away from the legacy codes in general,
all fencer API operations will be modified to return a full result.

This convenience function will come in handy for that.