Refactor: fencer: track full result for direct agent actions
This renames stonith_device_action() to execute_agent_action() for readability,
and has it set a full result rather than return a legacy return code.

As of this commit, handle_request() just maps the result back to a legacy code,
but it will make better use of it with planned changes.