Test: cts-cli: Change the resources in partially-active-group.
dummy-2 is now not running because it failed to start due to an
unimplemented feature.  I don't know what could possibly be
unimplemented about a dummy resource, but it's not important.

There is also a new dummy-3 resource that acts exactly the same as
dummy-2.  This preserves checking that the inactive member output can
still be displayed.

There are no code changes yet to properly handle displaying these
operations, so the results here just reflect the current handling.