Test: cts-cli: Add a failed probe on a new dummy-4 resource.
This is to verify that these resources which are part of a group are
displayed properly.  No code changes will be necessary, since groups are
just several other resources all in the same pile.

There are no code changes yet to properly handle displaying these
operations, so the results here just reflect the current handling.