Test: cts: Add --show-detail to a couple of the crm_mon tests.
This straightens out a couple differences in output between running
tests locally (where --enable-compat-2.0 is not given, which would
automatically add --show-detail) and running tests under mock (where
that option is given).

Note that this only really matters for failed resource actions, which
were not previously output as part of any crm_mon regression test.  It
is only the patches in this series that have introduced those, and thus
this difference.