Refactor: fencer: ensure all requests get clean-up
handle_request() has if-else blocks for each type of request. Previously, if a
request didn't need a reply, the function would do any clean-up needed and
return immediately. Now, we track whether a reply is needed, and all request
types flow to the end of the function for consistent clean-up.

This doesn't change any behavior at this point, but allows us to do more at the
end of request handling.