Test: cts-cli: Update the last-rc-change sed expression.
This can now occur in both the XML output (where it's wrapped in double
quotes) and the text output (where it's wrapped in single quotes and
followed by a comma).  In addition, a plus or minus can occur in the
time string.

The "{0,1}" syntax takes the place of a "?" for marking the optional
comma.  In FreeBSD sed, "?" doesn't mean anything special.