Low: fencer: handle topology level registration errors better
Rename stonith_level_register() to fenced_register_level() for consistency, and
refactor it to return a full result rather than a legacy return code.

Return a protocol error for missing information in the request XML, and log
invalid level numbers at warning level. Use a new combination of
PCMK_EXEC_INVALID with CRM_EX_INVALID_PARAM for invalid levels, so it gets
mapped back to the legacy code -EINVAL (which was returned before).