Low: fencer: handle topology level unregistration errors better
Rename stonith_level_remove() to fenced_unregister_level() for consistency, and
refactor it to return a full result rather than a legacy return code.

Return a protocol error for missing information in the request XML, and log
invalid level numbers at warning level. Use PCMK_EXEC_INVALID with
CRM_EX_INVALID_PARAM for invalid levels, so it gets mapped back to the legacy
code -EINVAL (which reverses the recent change in ec60f014b, both for backward
compatibility and because it makes sense -- a missing parameter is a protocol
error, while an invalid parameter is an invalid parameter error).