Low: libcrmservice: Handle systemd service templates.
These unit files (which have an @ sign at the end) expect to be
parameterized by an instance name.  Not providing an instance name
causes the dbus lookup to fail, and we fall back to assume this is an
LSB service.  If the user doesn't provide an instance name, just add a
fake one.  It doesn't seem to matter what name is given for the lookup.

See: rhbz#2003151