Feature: controller,tools: improve description for fencing alerts/traps
This functionizes creating a description for fencing events, so it can be used
by both the controller for alerts and crm_mon for traps, for consistency.

Now that we have the full result including exit reason, we can improve the
description, but the format is kept similar to before to minimize the change.

The alert/trap also includes the legacy return code for the event, but we can't
change that now because lrmd_send_fencing_alert() and the alert/trap
environment variables are public API.