Refactor: libpacemaker: tweak the sense of should_add_action_to_graph()
Previously, should_add_action_to_graph() returned false if the action was
already added to the graph. However that is nonintuitive, and two of the three
callers explicitly checked that condition first so it wouldn't hit.

Now, should_add_action_to_graph() does not check whether the action was already
added, and instead the lone caller that needs that check does it directly.