Test: cts-fence-helper: update expected return code
Before recent changes, libstonithd obtained the fence API's legacy result code
directly from the fencer's XML reply, meaning that the legacy code was the
result of the fencer's mapping of the full result (including the action stderr).

After those changes, libstonithd now ignores the legacy code in the fencer's
reply, and instead maps the legacy code itself from the full result in the
fencer's reply.

However, the fencer's reply does not have the action stderr, so failures that
mapped to -pcmk_err_generic on the server side now map to -ENODATA on the
client side. Update cts-fence-helper's expected return code to match (neither
code is particularly useful, so there wouldn't be much benefit from having the
fencer pass the action stderr with replies, which would be considerable
additional work).