Refactor: controller,scheduler: remove vestigial probe_complete handling
8f76b7821 (1.1.14) stopped generating probe_complete actions in the
scheduler, but much of the code handling it was left intact. Since the
scheduler is only ever run locally, we don't have to worry about backward
compatibility.

We do need to keep the probe_complete node attribute, because the attribute
manager learns a node is remote when the attribute is cleared when the node
comes up, and because older DCs in a rolling upgrade and old saved CIBs might
use it.