Refactor: libcrmcommon: drop unnecessary system error redefinitions
portability.h defines some system error codes that might not be present on
non-Linux systems.

This was a bad idea, since there's no way to ensure the defined values don't
conflict with existing system codes. However, we use a number of them, so it's
probably best to keep them, at least until we can make a backward compatibility
break.

However, we don't use EUNATCH, ENOSR, or ENOSTR, so we can delete those.