Log: libcrmcommon: improve handling of portability.h error codes
portability.h defines some system error codes that might not be present on
non-Linux systems.

Define a constant for each one (for example, PCMK__ECOMM for ECOMM) when
the system doesn't have the value, so we can detect that when relevant.

Also, make sure pcmk_rc_name() and pcmk_rc_str() handle all of these values.