Refactor: scheduler: simplify copying action meta-attributes
Rename dup_attr() to copy_meta_to_notify() for readability, and add a doxygen
block.

Also, reimplement it to take the notify action as the user data instead of the
notify action's meta-attribute table. This makes the semantics more obvious,
and we can avoid using add_hash_param() since we don't need everything it does.