Log: scheduler: improve messages for action configuration changes
Previously, if an action configuration changed, its XML configuration would be
logged at info level, which seems a bit chatty. Lower it to debug.

Also, if a resource needed to be reloaded, but the agent didn't support
reloads, or a recurring monitor needed to be rescheduled, the logs would still
say "reload". Use a more appropriate message instead.

To make this easier, and for readability and to reduce code duplication,
functionize forcing a restart due to an action configuration change.