Refactor: libcrmcommon: Rename pcmk__print_output_metadata...
to pcmk__format_output_metadata to make it more clear what it now does.
It also returns the metadata string that used to be printed.  This is so
that the output can be directed to various places.  It also means the
caller is responsible for freeing the string when they're done with it.