Refactor: tools: Update output format handling in mon_refresh_display.
If the output format is nothing useful, just return.  There's no point
in doing everything else in the function just to throw it away.

Then, get rid of the error handling if writing the output HTML file
fails.  This wouldn't get run anyway.  The cluster-status output message
always returns pcmk_rc_ok, and the error this block is trying to handle
would actually happen when pcmk__output_new is called.

With that done, we're basically just deciding between printing the
simple monitor format or any of the more complex output formats, all of
which can be handled by the same function call.