Doc: libcrmservice: Clarify return values
services_action_async_fork_notify():

The existing comment was incorrect. TRUE doesn't necessarily mean
successful initiation of execution.

execute_action() and services__execute_file():

I first interpreted the previous version as "because it already was
pending or is pending," rather than "because it already was freed or is
pending." This resolves the ambiguity.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>