Low: tools: Handle output format arguments before doing daemonize mode.
This allows us to catch the case where something nonsensical like
"crm_mon -d --output-to=foo" gets run.  In this case, crm_mon opens the
output file, forks, and returns to the command line but nothing ever
appears in the output file because output_format is mon_output_none.

Instead, move all the daemonize code to after reconcile_output_args and
add_output_args have been run and the output object created.  It is
only after the first two have been run that we can trust that
output_format will be correct and will not be changed (hence the
comment, which I have also moved).