Refactor: scheduler: allocate and free data set every request
The memory allocation of a pe_working_set_t is trivial compared to all the
allocations done as part of the scheduling, not to mention it was reset for
every request, so there's no real benefit to it.

Refactor init_working_set() to always return a freshly allocated working set,
and have handle_pecalc_op() free it when done processing the request.