Test: libcrmcommon: Add unit tests for compare_version.
Note that there are no tests here for comparing version strings that
contain anything aside from numbers.  compare_version is fairly simple
and does not take those kinds of strings into account.  It's easy to
hang it by passing it something like "1.0a".  On the other hand, it
appears to only be used for comparing things like CIB versions, which
don't come in that format.