Refactor: libcrmcommon: The IPC dispatch function should return a bool.
This can be used to determine whether more reply messages should be
expected after a given one is handled.

However, just this bit of information is not enough.  Currently, more
responses follow an ACK only in two cases - the ping and quit messages
for the pacemakerd daemon.  In the dispatch function, the only way to
tell these two cases apart from all others is if we also check status
for CRM_EX_INDETERMINATE.