Refactor: libcrmcommon: Read IPC replies as long as there are more.
The replies to some IPC messages (most notably the ping message for
pacemakerd) return an ack followed by the actual reply.  We need to get
the ack and then continue reading and dispatching messages so that the
actual reply is handled.  This is done by calling crm_ipc_read
repeatedly until it no longer returns anything.