Test: scheduler: Add unit tests for native_find_rsc.
Note that there's some fairly cheesy resource finding loop code in here.
This is because we can't use the existing resource finding functions,
because that is what we are testing with this patch.

Finally, note that this requires cts/cli/crm_mon.xml.  If that file
changes, this test case will have to be changed too.