Fix: fencing: delegate shouldn't matter when checking equivalent fencing
If a fencing action fails, but a later equivalent action succeeds, crm_mon will
show "(a later attempt succeeded)" when displaying the failed action.

Previously, an action was considered equivalent if it had the same target,
action, and delegate. However the delegate (i.e. the node executing the fence
action) doesn't really matter for this purpose.