Low: tools: Always output to the HTML dest file.
Previously, if you did --output-as=html --output-to= and backgrounded
the crm_mon process, it would not print anything to the destination file
until crm_mon exited.  However, if you added --daemonize, it would
output immediately.  This patch just makes the background behavior the
same as the daemonize behavior.

See: clbz#5492