Med: libpacemaker: Fix a segfault in crm_mon.
If connecting to the CIB fails (for instance, CIB_file=nonexistent.xml),
do not clean up the connection.  The caller supplied the connection, so
it is up to the client to clean up the connection.  In crm_mon's case,
it was already cleaning up, leading to a double free and therefore a
segfault.