Feature: tools: Finish using formatted output in crm_attribute.
This includes the introduction of an "attribute" message, which does
partially duplicate node-attribute output message used in crm_mon.
However, the output of these two tools needs to be maintained for anyone
who might be relying on the output for scripting.  Thus, the new message
is only in crm_attribute, and not in any library.