Low: fencer: always add device ID to level addition notification
When a topology level addition request comes into the fencer,
handle_level_add_request() processes it and sends a notification with the
result.

The notification also includes a level description based on the request, but
the description was not created for every possible error case. It was only set
if fenced_register_level() was called, not for unprivileged requests, and
wasn't even set by that function on early errors.

Now, handle_level_add_request() creates the description for the unprivileged
case. Part of fenced_register_level() was functionized to allow this.