Low: fencing: add device/level description to configuration change notifications
The fencer sends a notification for adding or removing a device or topology
level. Previously, this contained the result and nothing else. Now, add the
the description of what was added or removed.

The only place this matters currently is in cts-fence-helper log messages,
but it seems like a good idea.