Low: fencer: avoid redundant result in XML
Client notifications for fencing actions and peer broadcasts of fencing results
both use fencing_result2xml() to create the data subsection of the message XML.

Client notifications set and get the result from the top-level XML (via
fenced_send_notification() and libstonithd's xml_to_event()), but peer
broadcasts set and get the result from the data subsection.

Previously, fencing_result2xml() added the result for peer broadcasts, but this
was redundant and unused for client notifications. Now, add the result when
creating the peer broadcast instead. This wasn't as big a deal when the result
was just an integer legacy return code, but the full result can include stdout
and stderr.