Low: tools: Fix one method of restarting a cloned group.
If given a cloned group resource and a node in crm_resource, we need to
restart the resource by banning it from the node.  Previously, this was
not working because we were attempting to ban the name of the cloned
group.  Instead, we need to ban the name of the clone itself.