Low: resources: Fix a bug in init_smart in HealthSMART.
Previously, this loop would iterate over everything in $DRIVES and for
each of those, it would iterate over everything in $OCF_RESKEY_devices
and then run smartctl for every combination.  This is probably not what
was ever expected.

Instead, those two lists are assumed to be the same length and for each
index, $DRIVES[i] has the type of $OCF_RESKEY_devices[i].

And while I'm at it, set the $OCF_RESKEY_drives default in the top
block, just like all the other parameters.