High: tools: Fix the -v option to attrd_updater.
Two things were going wrong here:

(1) The "-v" option needs to be recognized by the command_cb function.

(2) All options must have a long name, so duplicate the "update" name.
Without a long name, the leading NULL is seen as the terminator for the
entire entry array and neither that entry nor any after it will be added
to the program.

Fixes T195.