Feature: scheduler: allow-unhealthy-nodes resource meta-attribute
One of the pain points with the node health feature is that a strategy like
migrate-on-red will ban the health monitors themselves from an unhealthy node,
meaning that the cluster can never detect when the node becomes healthy again.
Manual clearing of any health attributes is required.

Now, resources may have a new meta-attribute "allow-unhealthy-nodes" which
exempts the resource from negative health scores.

This establishes a new convention that constants for meta-attribute names
will be defined in include/crm/common/options_internal.h with the prefix
"PCMK__META_".

Fixes T24