Feature: libpacemaker: Add host to the attribute message.
And then change the output string on the text formatter to be in line
with what attrd_updater would display.  Note that attrd_updater does not
display scope or instance, but we can take care of that by just passing
in NULL for those things.

Also note that this breaks the output format for crm_attribute.  That
fix is coming in a future patch.