Refactor: all: use constants for node-health-related option values
This establishes a new convention that constants for enumerated option values
will be defined in include/crm/common/options_internal.h with the prefix
"PCMK__VALUE_".

These constants are not option-specific; for example, PCMK__VALUE_NONE is
defined as "none" and can be used with multiple options.

The constants are used only where they are magic strings used as values for
Pacemaker option values, not where they are used as regular words or as values
for external software options (like dot file colors).