Refactor: libcrmcommon: add enum and parser for node health strategy
This creates new health.c and health_internal.h files in libcrmcommon for code
related to node health. Currently all such code is in the scheduler libraries,
but some of that will eventually be moved here.

For now, there is a new enum pcmk__health_strategy and a function
pcmk__parse_health_strategy() to parse it from a string. Nothing uses it as of
this commit.